+++ samples/runGangnam_v2.sh
@@ -8,4 +8,4 @@ if [ ! -f "Psy-Gangnam_Style.mp4" ]; then
 	yt-dlp -f 18 http://youtu.be/9bZkp7q19f0 -o Psy-Gangnam_Style.mp4
 fi
 
-../emwy.py -y gangnam.emwy.yaml
+../emwy_cli.py -y gangnam.emwy.yaml